Updates for extension

+++ CurrentRangeUpdate.h
@@ -5,6 +5,6 @@ int* Sort_CurrentValues(int* CurrentValues, int samples);
 void PrintonConsole(int CurrentRange_Count);
 
 int BinaryToDecimalConverter(int A2D12bit);
-int* ConvertCurrentValuestoDecimal(unsigned long int* CurrentValues, int samples);
-int CurrentRangeUpdate_Binary (unsigned long int* CurrentValues, int samples);
+int* ConvertCurrentValuestoDecimal(unsigned long long int* CurrentValues, int samples);
+int CurrentRangeUpdate_Binary (unsigned long long int* CurrentValues, int samples);
 int A2DinputScaler(int currentvalue);
